requirements.txt
pyyaml>=5.4.1
pathlib>=1.0.1
typing-extensions>=4.0.1 